Handling for fcn appl, except, maybeBound opApplication
and generally improves the some of the handlings as well.

I've improved the test suite for the large specs, by applying
the formatter again to the output of the first formatter..
This helped fixing a lot of issues and helps ensuring the produced
specs are correct and formatter's output is stable.